AG-9963 Move axis validation into prepareAxis.ts
