+++ packages/ag-charts-community/src/chart/axis/axisUtil.ts
@@ -1,15 +1,8 @@
 import { FROM_TO_MIXINS } from '../../motion/fromToMotion';
 import type { FromToMotionPropFn, NodeUpdateState } from '../../motion/fromToMotion';
-import type {
-    AgBaseAxisOptions,
-    AgBaseCartesianAxisOptions,
-    AgCartesianAxisPosition,
-    AgCartesianAxisType,
-} from '../../options/agChartOptions';
 import type { Group } from '../../scene/group';
 import type { Line } from '../../scene/shape/line';
 import type { Text } from '../../scene/shape/text';
-import type { ChartAxis } from '../chartAxis';
 
 export type AxisLineDatum = { x: number; y1: number; y2: number };
 export type AxisDatum = {
@@ -218,62 +211,4 @@ export function resetAxisLineSelectionFn() {
     return (_node: Line, datum: AxisLineDatum) => {
         return { ...datum };
     };
-}
-
-const CARTESIAN_AXIS_POSITIONS: AgCartesianAxisPosition[] = ['top', 'right', 'bottom', 'left'];
-const CARTESIAN_AXIS_TYPES: AgCartesianAxisType[] = ['category', 'grouped-category', 'number', 'log', 'time'];
-
-function hasCartesianAxisPosition(axis: ChartAxis): axis is ChartAxis & { position: AgCartesianAxisPosition } {
-    const allowedTypes: string[] = CARTESIAN_AXIS_TYPES;
-    return allowedTypes.includes(axis.type);
-}
-
-function isCartesianAxisOptions(options: AgBaseAxisOptions): options is AgBaseCartesianAxisOptions {
-    const allowedTypes: string[] = CARTESIAN_AXIS_TYPES;
-    return allowedTypes.includes(options.type);
-}
-
-function isAxisPosition(position: unknown): position is AgCartesianAxisPosition {
-    const allowedPositions: string[] = CARTESIAN_AXIS_POSITIONS;
-    return typeof position === 'string' && allowedPositions.includes(position);
-}
-
-// If axis[].position, then we cannot always default to the same value. We need
-// to default to an 'untaken' position (see AG-9963 for more info).
-export class AxisPositionGuesser {
-    private result: ChartAxis[] = [];
-    private valid: ChartAxis[] = [];
-    private invalid: ChartAxis[] = [];
-
-    push(axis: ChartAxis, options: AgBaseAxisOptions) {
-        const { result, valid, invalid } = this;
-        if (isCartesianAxisOptions(options)) {
-            if (isAxisPosition(options.position)) {
-                valid.push(axis);
-            } else {
-                invalid.push(axis);
-            }
-        }
-
-        result.push(axis);
-    }
-
-    guessInvalidPositions() {
-        const takenPosition: (AgCartesianAxisPosition | undefined)[] = this.valid
-            .filter((v) => hasCartesianAxisPosition(v))
-            .map((v) => v.position)
-            .filter((v) => v !== undefined);
-        const guesses: AgCartesianAxisPosition[] = ['top', 'right', 'bottom', 'left'];
-        for (const invalidAxis of this.invalid) {
-            let nextGuess = guesses.pop();
-            while (takenPosition.includes(nextGuess) && nextGuess !== undefined) {
-                nextGuess = guesses.pop();
-            }
-
-            if (nextGuess === undefined) break;
-            invalidAxis.position = nextGuess;
-        }
-
-        return this.result;
-    }
 }
